refactor(#118): slice 5 — extract TypeInference partial class
Move ValidateExpression, TryInferKind, TryInferFunctionCallKind, and
TryInferBinaryKind from PreceptTypeChecker.cs into the new partial-class
file PreceptTypeChecker.TypeInference.cs (~748 lines).

Pure structural refactor: no behavior changes, no visibility changes,
no signature changes, no caller updates. Cross-file calls into Helpers,
Narrowing, and ProofChecks preserved unchanged.

Validation: dotnet build (0 new warnings), 1752/1752 Precept tests pass
including the Slice 5 anchor set (PreceptTypeCheckerTests,
ConditionalExpressionTests, StringAccessorTests: 211/211).
